[fix]雷鳴の審判のスキルを一つに統合
