+++ Assets/Data/SkillData/LightningStrike 1.asset
@@ -9,12 +9,12 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c5892dca78e5fe94ea0db661a00e2312, type: 3}
-  m_Name: LightningStrike
+  m_Script: {fileID: 11500000, guid: be655b64f5d7b3f44bc4e975e4b5b069, type: 3}
+  m_Name: LightningStrike 1
   m_EditorClassIdentifier: Assembly-CSharp::LightningStrike
   _id: 666
   _priority: 0
-  _name: "\u96F7\u9CF4\u306E\u5BE9\u5224"
+  _name: "\u96F7\u9CF4\u306E\u5BE9\u52241"
   _explanation: 
   _icon: {fileID: 0}
   _timing: 2
@@ -22,4 +22,6 @@ MonoBehaviour:
   _maxInterval: 4
   _damegeMultiplier: 1
   _searchRadius: 999
-  _hitEffectPrefab: {fileID: 8199528857376779805, guid: 7b0b065706e852543bdbebf0e9401ff7, type: 3}
+  _hitEffectPrefabs: []
+  _areaRadius: 0
+  _electricShockDuration: 0
