+++ Assets/Data/SkillData/LightningStrike 1.asset.meta
