+++ Assets/Data/SkillData/LightningStrike 2.asset
@@ -1,0 +1,27 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: be655b64f5d7b3f44bc4e975e4b5b069, type: 3}
+  m_Name: LightningStrike 2
+  m_EditorClassIdentifier: Assembly-CSharp::LightningStrike
+  _id: 0
+  _priority: 0
+  _name: 
+  _explanation: 
+  _icon: {fileID: 0}
+  _timing: 0
+  _minInterval: 2
+  _maxInterval: 4
+  _damegeMultiplier: 1.6
+  _searchRadius: 999
+  _hitEffectPrefabs: []
+  _areaRadius: 0
+  _electricShockDuration: 0
