+++ Assets/Data/SkillData/LightningStrike 2.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 074a648368c418547b1e598b926d4cd1
+guid: a05c65a211dbe2d498e4beba93a2dae0
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
