+++ Assets/Data/SkillData/LightningStrike 3.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 53722a74beee9cd4e82e6941a3cd16e7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
