--- Assets/Data/SkillData/LightningStrike2.asset
@@ -1,27 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1f5dad3e1dd965d4089f2a62e818fb24, type: 3}
-  m_Name: LightningStrike2
-  m_EditorClassIdentifier: Assembly-CSharp::LightningStrike2
-  _id: 667
-  _priority: 0
-  _name: "\u96F7\u9CF4\u306E\u5BE9\u52242"
-  _explanation: 
-  _icon: {fileID: 0}
-  _timing: 2
-  _minInterval: 2
-  _maxInterval: 4
-  _damegeMultiplier: 1.3
-  _searchRadius: 999
-  _hitEffectPrefabs:
-  - {fileID: 8199528857376779805, guid: 7b0b065706e852543bdbebf0e9401ff7, type: 3}
-  - {fileID: 240071800827797065, guid: 9a23160c45ecee049a37b129d8746d16, type: 3}
