+++ Assets/Scripts/Data/Skill/LightningStrike.cs
@@ -1,7 +1,8 @@
+using System.Collections.Generic;
 using UnityEngine;
 
 [CreateAssetMenu(fileName = "LightningStrike", menuName = "GameData/Skill/LightningStrike")]
-public class LightningStrike : SkillBase, ISkillUpdater
+public class LightningStrike : SkillBase,ISkillUpdater
 {
     public void OnUpdate(float deltaTime, PlayerMode mode, IPlayerStats stats, Vector3 playerPosition, EnemyManager enemyManager)
     {
@@ -14,45 +15,80 @@ public class LightningStrike : SkillBase, ISkillUpdater
         _timer = Random.Range(_minInterval, _maxInterval);
 
         Trigger(stats, playerPosition, enemyManager);
-        Debug.Log("落雷発動");
     }
 
+    [SerializeField] private int _targetCount = 1;
     [SerializeField] private float _minInterval = 2f;
     [SerializeField] private float _maxInterval = 4f;
-    [SerializeField] private float _damegeMultiplier = 1f;
+    [SerializeField] private float _damageMultiplier = 1.6f;
     [SerializeField] private float _searchRadius = 999f;
-    [SerializeField] private GameObject _hitEffectPrefab;
+    //エフェクト
+    [SerializeField] private GameObject[] _hitEffectPrefabs;
+    //範囲
+    [SerializeField] private float _areaRadius;
+    //感電時間
+    [SerializeField] private float _electricShockDuration;
 
     private float _timer;
 
-    private void Trigger(IPlayerStats stats,Vector3 playerPosition, EnemyManager enemyManager)
+    private void Trigger(IPlayerStats stats, Vector3 playerPosition, EnemyManager enemyManager)
     {
-        var enemies = enemyManager.GetEnemiesInRange(playerPosition,_searchRadius);
+        var enemies = enemyManager.GetEnemiesInRange(playerPosition, _searchRadius);
 
-        Debug.Log($"敵の数: {enemies.Count}");
+        if (enemies.Count == 0) return;
 
-        if (enemies.Count == 0f) return;
+        var shuffled = new List<IEnemy>(enemies);
+        Shuffle(shuffled);
 
-        var target = enemies[Random.Range(0,enemies.Count)];
+        int targetCount = Mathf.Min(_targetCount, shuffled.Count);
 
-        Debug.Log($"ターゲット: {target}");
+        var effects = new List<GameObject>(_hitEffectPrefabs);
+        Shuffle(effects);
 
-        float damage = stats.AttackPower * _damegeMultiplier;
-
-        var damageContext = new DamageContext { AttackPower = damage, PlayerMode = PlayerMode.Thunder };
-
-        target.TakeDamage(damageContext);
-
-        //仮のヒットエフェクト
-        if(_hitEffectPrefab != null)
+        for (int i = 0; i < targetCount; i++)
         {
-            var effect = Instantiate(_hitEffectPrefab, target.Position, Quaternion.identity);
-            Destroy(effect, 2f);
-        }
-        
-        Debug.Log("落雷発動");
+            var target = shuffled[i];
 
-        // 視覚確認
-        Debug.DrawLine(playerPosition, target.Position, Color.blue, 1f);
+            var areaTragets = enemyManager.GetEnemiesInRange(target.Position, _areaRadius);
+
+            foreach (var enemy in areaTragets)
+            {
+                float damage = stats.AttackPower * _damageMultiplier;
+
+                var damageContext = new DamageContext { AttackPower = damage, PlayerMode = PlayerMode.Thunder };
+                enemy.TakeDamage(damageContext);
+
+                enemy.ConditionController?.ApplyCondition(new ElectrifiedCondition(_electricShockDuration, enemyIsBoss: false));
+            }
+
+            GameObject effectPrefab = null;
+            if (effects.Count > 0)
+            {
+                effectPrefab = effects[i % effects.Count];
+            }
+
+            //仮のヒットエフェクト
+            if (effectPrefab != null)
+            {
+                var effect = Instantiate(effectPrefab, target.Position, Quaternion.identity);
+                Destroy(effect, 2f);
+            }
+            // 視覚確認
+            Debug.DrawLine(playerPosition, target.Position, Color.blue, 1f);
+        }
+        Debug.Log("落雷発動");
+    }
+    /// <summary>
+    /// エフェクトシャッフル
+    /// </summary>
+    /// <typeparam name="T"></typeparam>
+    /// <param name="list"></param>
+    private void Shuffle<T>(List<T> list)
+    {
+        for (int i = list.Count - 1; i > 0; i--)
+        {
+            int rand = Random.Range(0, i + 1);
+            (list[i], list[rand]) = (list[rand], list[i]);
+        }
     }
 }
