+++ Assets/Scripts/Data/Skill/LightningStrike.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: c5892dca78e5fe94ea0db661a00e2312
+guid: be655b64f5d7b3f44bc4e975e4b5b069
