--- Assets/Scripts/Data/Skill/LightningStrike2.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1f5dad3e1dd965d4089f2a62e818fb24
